Create new .sort sub package
