+++ app/src/main/java/com/poupa/vinylmusicplayer/discog/Discography.java
@@ -9,13 +9,13 @@ import androidx.annotation.Nullable;
 
 import com.poupa.vinylmusicplayer.App;
 import com.poupa.vinylmusicplayer.discog.tagging.TagExtractor;
-import com.poupa.vinylmusicplayer.helper.SortOrder;
 import com.poupa.vinylmusicplayer.interfaces.MusicServiceEventListener;
 import com.poupa.vinylmusicplayer.model.Album;
 import com.poupa.vinylmusicplayer.model.Artist;
 import com.poupa.vinylmusicplayer.model.Genre;
 import com.poupa.vinylmusicplayer.model.Song;
 import com.poupa.vinylmusicplayer.provider.BlacklistStore;
+import com.poupa.vinylmusicplayer.sort.SongSortOrder;
 import com.poupa.vinylmusicplayer.ui.activities.MainActivity;
 import com.poupa.vinylmusicplayer.util.FileUtil;
 import com.poupa.vinylmusicplayer.util.PreferenceUtil;
@@ -199,7 +199,7 @@ public class Discography implements MusicServiceEventListener {
             }
         }
         // Maintain sorted album after merge
-        Collections.sort(fullAlbum.songs, SortOrder.BySong.BY_DISC_TRACK);
+        Collections.sort(fullAlbum.songs, SongSortOrder.BY_DISC_TRACK);
         return fullAlbum;
     }
