+++ app/src/main/java/com/poupa/vinylmusicplayer/discog/MemCache.java
@@ -2,12 +2,12 @@ package com.poupa.vinylmusicplayer.discog;
 
 import androidx.annotation.NonNull;
 
-import com.poupa.vinylmusicplayer.helper.SortOrder;
-import com.poupa.vinylmusicplayer.loader.SongLoader;
 import com.poupa.vinylmusicplayer.model.Album;
 import com.poupa.vinylmusicplayer.model.Artist;
 import com.poupa.vinylmusicplayer.model.Genre;
 import com.poupa.vinylmusicplayer.model.Song;
+import com.poupa.vinylmusicplayer.sort.AlbumSortOrder;
+import com.poupa.vinylmusicplayer.sort.SongSortOrder;
 
 import java.util.ArrayList;
 import java.util.Collections;
@@ -61,10 +61,10 @@ class MemCache {
         for (Long artistId : albums.keySet()) {
             Artist artist = artistsById.get(artistId);
             if (artist == null) continue;
-            Collections.sort(artist.albums, SortOrder.ByAlbum.BY_YEAR_DESC);
+            Collections.sort(artist.albums, AlbumSortOrder.BY_YEAR_DESC);
         }
         for (Album album : albums.values()) {
-            Collections.sort(album.songs, SortOrder.BySong.BY_DISC_TRACK);
+            Collections.sort(album.songs, SongSortOrder.BY_DISC_TRACK);
         }
     }
