+++ app/src/main/java/com/poupa/vinylmusicplayer/loader/GenreLoader.java
@@ -3,10 +3,10 @@ package com.poupa.vinylmusicplayer.loader;
 import androidx.annotation.NonNull;
 
 import com.poupa.vinylmusicplayer.discog.Discography;
-import com.poupa.vinylmusicplayer.helper.SortOrder;
-import com.poupa.vinylmusicplayer.util.StringUtil;
 import com.poupa.vinylmusicplayer.model.Genre;
 import com.poupa.vinylmusicplayer.model.Song;
+import com.poupa.vinylmusicplayer.sort.SongSortOrder;
+import com.poupa.vinylmusicplayer.util.StringUtil;
 
 import java.util.ArrayList;
 import java.util.Collection;
@@ -28,7 +28,7 @@ public class GenreLoader {
         }
         else {
             ArrayList<Song> songs = new ArrayList<>(genreSongs);
-            Collections.sort(songs, SortOrder.BySong.BY_ALBUM);
+            Collections.sort(songs, SongSortOrder.BY_ALBUM);
             return songs;
         }
     }
